gemini2/requirements.txt
langchain_core==0.3.65
langchain_google_genai==2.1.5
Pillow==9.0.1
python-dotenv==1.1.0
streamlit==1.45.1
flask